refactor: extract tailwind preset check
