+++ scripts/check-tailwind-preset.ts
@@ -1,0 +1,26 @@
+import { createRequire } from "node:module";
+import preset from "../packages/tailwind-config/src/index.ts";
+
+const require = createRequire(import.meta.url);
+
+let resolvedPresetPath = "<unresolved>";
+try {
+  resolvedPresetPath = require.resolve("@acme/tailwind-config");
+  // eslint-disable-next-line no-console
+  console.log(
+    `[tailwind.config] ✅  @acme/tailwind-config resolved → ${resolvedPresetPath}`
+  );
+} catch (err) {
+  // eslint-disable-next-line no-console
+  console.error(
+    "[tailwind.config] ❌  @acme/tailwind-config could NOT be resolved.\n" +
+      "Is the package in pnpm-workspace.yaml? Did you run `pnpm install`?",
+    err
+  );
+}
+
+// eslint-disable-next-line no-console
+console.log(
+  "[tailwind.config] ℹ️  preset keys:",
+  preset && typeof preset === "object" ? Object.keys(preset) : "not an object"
+);
